+++ internal/cmd/templates.go
@@ -51,6 +51,7 @@ var (
 			flags.bindJSON(flagSet)
 			flags.bindFix(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -88,6 +89,7 @@ var (
 			flags.bindDryRun(flagSet)
 			flags.bindJSON(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -225,6 +227,7 @@ If Vim integration is set up, files will be generated when you open a new Protob
 			flags.bindOverwrite(flagSet)
 			flags.bindFix(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -240,6 +243,7 @@ If Vim integration is set up, files will be generated when you open a new Protob
 			flags.bindDryRun(flagSet)
 			flags.bindJSON(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -329,6 +333,7 @@ $ cat input.json | prototool grpc example \
 			flags.bindMethod(flagSet)
 			flags.bindStdin(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -371,6 +376,7 @@ $ cat input.json | prototool grpc example \
 			flags.bindListAllLinters(flagSet)
 			flags.bindListLinters(flagSet)
 			flags.bindProtocURL(flagSet)
+			flags.bindNoCache(flagSet)
 		},
 	}
 
@@ -506,6 +512,12 @@ func getRunner(stdin io.Reader, stdout io.Writer, stderr io.Writer, flags *flags
 		runnerOptions = append(
 			runnerOptions,
 			exec.RunnerWithJSON(),
+		)
+	}
+	if flags.noCache {
+		runnerOptions = append(
+			runnerOptions,
+			exec.RunnerWithNoCache(),
 		)
 	}
 	if flags.printFields != "" {