+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Accept `prototool.json` files for configuation in addition to
   `prototool.yaml` files.
 - Add `--config-data` flag.
+- Add a `no-cache` flag to disable protoc caching.
 
 
 ## [1.2.0] - 2018-08-29